Backup of state of 2018
For some reasons (e.g. missing CPU/RAM power) the project was delayed
by me, but not all changes are commited. The grammar_def.hpp was
enhanced with respect to the comments, but not completely finished
(e.g. \code{.bnf} ...)
IIRC the changes where correct, even incomplete, so I commit them
in 2021.